requirements.txt
numpy==2.1.1
pillow==10.4.0
PyYAML==6.0.2
requests==2.32.3
torch==2.4.1
torchvision==0.19.1
tqdm==4.66.5
